+++ esm/with-options.js
@@ -1,0 +1,5 @@
+export function withOptions (options = {}) {
+  return (requestParams) => {
+    return { ...requestParams, ...options };
+  };
+}